Move collector inside the binder
