+++ spring-cloud-stream-test-support/src/test/java/org/springframework/cloud/stream/test/ExampleTest.java
@@ -28,7 +28,7 @@ import org.springframework.boot.test.SpringApplicationConfiguration;
 import org.springframework.cloud.stream.annotation.EnableModule;
 import org.springframework.cloud.stream.annotation.ModuleChannels;
 import org.springframework.cloud.stream.annotation.Processor;
-import org.springframework.cloud.stream.test.binder.MessageCollector;
+import org.springframework.cloud.stream.test.binder.TestSupportBinder;
 import org.springframework.integration.annotation.Transformer;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.support.GenericMessage;
@@ -50,7 +50,7 @@ public class ExampleTest {
 	private Processor processor;
 
 	@Autowired
-	private MessageCollector messageCollector;
+	private TestSupportBinder.MessageCollectorImpl messageCollector;
 
 	@Test
 	@SuppressWarnings("unchecked")
